chore: create `.gitignore` file for Rust project
A .gitignore file was created to include common unneeded files in Rust projects. This aims to prevent tracking and commit of unnecessary files such as debug and target directories, Cargo.lock, backup files (*.rs.bk), and *.pdb files which are generated by MSVC Windows builds of rustc for debugging purposes.