chore: add `wrangler.toml` file for Worker Deployment
This commit introduces a new wrangler.toml file which details the name, main script and build command for Cloudflare Workers, enabling consistency in deployment across environments. The 'nounish-pro' in the name specifies the project, while the build command details how the project should be built for production by using cargo and worker-build. The compatibility date field is also set to ensure the script works with upcoming Workers runtime changes.