ci: add stale issue tracker file to Github
A '.github/stale.yml' file was added to enable automatic tracking and management of stale issues. Parameters such as 'daysUntilStale', 'daysUntilClose', and 'exemptLabels' were defined, as well as auto-comments for both marking and closing stale issues. This will assist in maintaining the repository by helping manage outdated or inactive issues.