refactor: implement URL parsing and transformation functionality
In order to ease the-URI-to-SQID transformation process, URL parsing functionality has been introduced to the 'main' event function. A 'post_async' block has been added to deal with the JSON payload that contains URLs, allowing parsing these incoming URLs easily.  The critical additions to facilitate this feature include Serde for payload serialization/deserialization and the URL crate. The logging level has been upgraded to 'debug' to facilitate detailed tracking and debugging of URL parsing processes. This enhancement ultimately improves the URL-to-SQID conversion efficiency and robustness, contributing to the overall reliability of the application.