refactor: add functionality to create OpenGraph images for titles
Updated the 'queries.rs' and 'utils.rs' to support creation of OpenGraph images for proposal and idea titles. For this, a new dependency 'urlencoding' has been added and a new function `create_og_image()` has been created in 'utils.rs'. This function generates a URL that points to an image rendered by Cloudinary, using the title and description of the proposal or idea. This image is then used as the OpenGraph image in proposal and idea methods in 'queries.rs'. This change will improve the visual representation of shared links on social platforms.