refactor: replace `urlencoding` with `percent_encoding` in utils
The urlencoding library was replaced with the percent_encoding library in utils.rs, to improve compatibility and stability. Global search and replace was performed on all its occurrences in the code.
Also, the text size is modified in the respective cloudinary url formats and some duplications and unnecessary parts are removed from the parts list. These changes were required for better code readability and performance.