feat: add unique color to each platform's og image
The color for the OG image is no longer hardcoded for every platform, but a unique color is now used for each one. This change allows for better visual differentiation of images associated with different platforms, improving user experience with our service and adding a touch of personalization. The code for creating an OG image was also modified in the utils file to accept this new color parameter and apply it when generating the image.