chore(deps): update lockfile version and package dependencies
The commit includes updates to the lockfile version as well as a reconfiguration and update of some package dependencies. This was done to ensure compatibility with newer versions of the dependencies and to improve overall performance.

Please note that this update will improve the project's maintainability and reduce potential security risks associated with using outdated packages.